Make code more maintainable, move uBPF frondend pass to ubpfModel
