+++ backends/ubpf/ubpfModel.h
@@ -95,8 +95,32 @@ namespace UBPF {
 
         static cstring reserved(cstring name) { return reservedPrefix + name; }
 
-        int numberOfParserArguments() { return version >= 20200515 ? 4 : 3; }
-        int numberOfControlBlockArguments() { return version >= 20200515 ? 3 : 2; }
+        int numberOfParserArguments() const { return version >= 20200515 ? 4 : 3; }
+        int numberOfControlBlockArguments() const { return version >= 20200515 ? 3 : 2; }
+
+        class getUBPFModelVersion : public Inspector {
+            bool preorder(const IR::Declaration_Constant *dc) override {
+                if (dc->name == "__ubpf_model_version") {
+                    auto val = dc->initializer->to<IR::Constant>();
+                    UBPFModel::instance.version = static_cast<unsigned>(val->value); }
+                return false; }
+            bool preorder(const IR::Declaration *) override { return false; }
+        };
+
+        const IR::P4Program *run(const IR::P4Program *program) {
+            if (program == nullptr)
+                return nullptr;
+
+            std::initializer_list<Visitor *> frontendPasses = {
+                    new getUBPFModelVersion,
+            };
+
+            PassManager passes(frontendPasses);
+            passes.setName("UBPFFrontEnd");
+            passes.setStopOnError(true);
+            const IR::P4Program *result = program->apply(passes);
+            return result;
+        }
     };
 
 }
